+++ wijisqs/broker.py
@@ -11,15 +11,12 @@ import concurrent
 import botocore.config
 import botocore.session
 
-
+from . import utils
 from . import buffer
 
 
 # See SQS limits: https://docs.aws.amazon.com/AWSSimpleQueueService/latest/SQSDeveloperGuide/sqs-limits.html
 # TODO: we need to add this limits as validations to this broker
-
-
-SQS_STATE: typing.Dict[str, bool] = {"QUEUE_CREATED": False, "TAGS_ADDED": False}
 
 
 class SqsBroker(wiji.broker.BaseBroker):
@@ -318,15 +315,14 @@ class SqsBroker(wiji.broker.BaseBroker):
         with concurrent.futures.ThreadPoolExecutor(
             thread_name_prefix=self._thread_name_prefix
         ) as executor:
-            if not SQS_STATE.get("QUEUE_CREATED"):
-                await self._get_loop().run_in_executor(
-                    executor, functools.partial(self._create_queue, queue_name=queue_name)
-                )
-            if not SQS_STATE.get("TAGS_ADDED"):
-                await self._get_loop().run_in_executor(
-                    executor, functools.partial(self._tag_queue, queue_name=queue_name)
-                )
+            await self._get_loop().run_in_executor(
+                executor, functools.partial(self._create_queue, queue_name=queue_name)
+            )
+            await self._get_loop().run_in_executor(
+                executor, functools.partial(self._tag_queue, queue_name=queue_name)
+            )
 
+    @utils.execute_only_once
     def _create_queue(self, queue_name: str) -> None:
         try:
             response = self.client.create_queue(
@@ -344,15 +340,14 @@ class SqsBroker(wiji.broker.BaseBroker):
             response.update({"event": "wijisqs.SqsBroker._create_queue", "queue_name": queue_name})
             self.logger.log(logging.DEBUG, response)
             self.QueueUrl = response["QueueUrl"]
-            SQS_STATE["QUEUE_CREATED"] = True
         except Exception as e:
             raise e
 
+    @utils.execute_only_once
     def _tag_queue(self, queue_name: str):
         response = self.client.tag_queue(QueueUrl=self.QueueUrl, Tags=self.queue_tags)
         response.update({"event": "wijisqs.SqsBroker._tag_queue", "queue_name": queue_name})
         self.logger.log(logging.DEBUG, response)
-        SQS_STATE["TAGS_ADDED"] = True
 
     async def enqueue(
         self, item: str, queue_name: str, task_options: wiji.task.TaskOptions